+++ utils/sessionClaimValidatorStore/index.js
@@ -2,16 +2,27 @@
  *
  * This software is licensed under the Apache License, Version 2.0 (the
  * "License") as published by the Apache Software Foundation.
- *
+ * 
  * You may not use this file except in compliance with the License. You may
  * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations
- * under the License.
- */
-export const package_version = "13.0.2";
-
-export const supported_fdi = ["1.8", "1.9", "1.10", "1.11", "1.12", "1.13", "1.14", "1.15"];
+ * under the License. 
+*/
+"use strict";
+function __export(m) {
+    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+exports.__esModule = true;
+let d = require("../../lib/build/utils/sessionClaimValidatorStore");
+if (d.default !== undefined) {
+    __export(d);
+} else {
+    __export({
+        default: d,
+        ...d,
+    });
+}